Fix data race in WithFieldGenerators
Fix a data race in WithFieldGenerators by ensuring that any call to
WithFieldGenerators makes a copy of the slice associated with the
parent context. This ensures that the backing array between
WithFieldGenerators calls is not shared (except in the case that it's
called with an empty slice), meaning that its append call will no
longer result in possibly writing to extra capacity in the context's
slice.

The context for this is that a recent pull request, #3, introduced
a data race when it added the WithFieldGenerators function to the
package.

To explain what happens, the WithFieldGenerators function works by
appending to the slice of field generators associated with a context.
If the context's associated slice grows enough, append will eventually
give it more capacity than it needs to accomodate further appends. When
that happens, any use of WithFieldGenerators across multiple goroutines
with the same parent context will introduce a data race by writing to
the extra capacity of the slice's backing array. Besides being a data
race, this also means that any data generated by field generators is
unreliable and may be either wrong or cause the program to crash
(by nature of it reading function pointers out of the slice).

An added test, TestFieldGeneratorRace, can be used to test this by
reverting the changes to field_generator.go without removing the test.
A typical scenario for this would be HTTP requests inheriting a base
context and attaching their own data to the slice, and then possibly
spawning additional goroutines on top of that context.